Updated logic for precision and added unit tests

+++ Backend/CalculationsApi/Models/CalculationRequest.cs
@@ -2,8 +2,8 @@ using System.Text.Json.Serialization;
 
 public class CalculationRequest {
     [JsonPropertyName("probabilityA")]
-    public float ProbabilityA { get; set; }
+    public decimal ProbabilityA { get; set; }
     [JsonPropertyName("probabilityB")]
-    public float ProbabilityB { get; set; }
+    public decimal ProbabilityB { get; set; }
 }
